+++ backend/src/neuroagent/app/main.py
@@ -131,7 +131,7 @@ async def lifespan(fastapi_app: FastAPI) -> AsyncContextManager[None]:  # type: 
             # trigger dynamic tool generation - only done once - it is cached
             _ = fastapi_app.dependency_overrides.get(
                 get_mcp_tool_list, get_mcp_tool_list
-            )(mcp_client)
+            )(mcp_client, app_settings)
             fastapi_app.state.mcp_client = mcp_client
             yield